extract context provider classes to file
