+++ javascript/src/androidMain/kotlin/dev/icerock/moko/javascript/ContextProvider.kt
@@ -1,0 +1,10 @@
+package dev.icerock.moko.javascript
+
+internal interface ContextProvider {
+    fun getBool(name: String): Boolean
+    fun getDouble(name: String): Double
+
+    fun getString(name: String): String
+
+    fun getScript(): String
+}
